+++ lib/product_matching_v2/domain/product_match_evidence.dart
@@ -1,0 +1,36 @@
+enum ProductMatchDiscoverySource { catalogName, catalogAlias }
+
+class ProductMatchEvidence {
+  ProductMatchEvidence({
+    required this.normalizedQuery,
+    required this.normalizedCatalogText,
+    required Iterable<String> matchedTokens,
+    required this.exactNormalizedMatch,
+    required this.discoverySource,
+  }) : matchedTokens = List.unmodifiable(matchedTokens);
+
+  factory ProductMatchEvidence.fromJson(Map<String, Object?> json) =>
+      ProductMatchEvidence(
+        normalizedQuery: json['normalizedQuery']! as String,
+        normalizedCatalogText: json['normalizedCatalogText']! as String,
+        matchedTokens: (json['matchedTokens']! as List<Object?>).cast(),
+        exactNormalizedMatch: json['exactNormalizedMatch']! as bool,
+        discoverySource: ProductMatchDiscoverySource.values.byName(
+          json['discoverySource']! as String,
+        ),
+      );
+
+  final String normalizedQuery;
+  final String normalizedCatalogText;
+  final List<String> matchedTokens;
+  final bool exactNormalizedMatch;
+  final ProductMatchDiscoverySource discoverySource;
+
+  Map<String, Object> toJson() => {
+        'normalizedQuery': normalizedQuery,
+        'normalizedCatalogText': normalizedCatalogText,
+        'matchedTokens': matchedTokens,
+        'exactNormalizedMatch': exactNormalizedMatch,
+        'discoverySource': discoverySource.name,
+      };
+}